Refactor player scraping so that it is done in batches of 100
Previously, it was based on some example code for SteamKit with little to no modifications.

This changes it from a Queue to a List and makes use of the fact that the GetPlayers() API can handle a list of steamids and not just one at a time (SteamKit was using a function that only took one at a time, but was calling the bulk one with a single element under the hood).